:construction: Severe refactoring for grain break
Breaking grains is going to require a huge amount of refactoring
which will be good to take some of the hidden functions in the
web workers. It will also mean improved coverage.

I also fixed an off-by-one error I was having that caused issues
that were mostly silent before. This was in the function that
produced equally sized grains.